requirements.txt
absl-py==2.1.0
accelerate==0.28.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.3
alembic==1.13.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
av==12.0.0
basicsr==1.4.2
beartype==0.18.2
bitsandbytes==0.43.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.2
CoCa-pytorch==0.1.0
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.16.0
decord==0.6.0
diffusers==0.27.2
dill==0.3.7
docopt==0.6.2
einops==0.7.0
einops-exts==0.0.4
einx==0.2.0
exceptiongroup==1.2.0
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.3
fonttools==4.50.0
frozendict==2.4.1
frozenlist==1.4.1
fsspec==2023.9.2
ftfy==6.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gradio==4.14.0
gradio_client==0.8.0
gradio_imageslider==0.0.18
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
ImageHash==4.3.1
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
jsonpickle==3.0.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.3
lightning==2.2.1
lightning-utilities==0.11.2
lit==18.1.3
lmdb==1.4.1
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
munch==4.0.0
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.0.6
openai-clip==1.0.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
optuna==3.6.1
orjson==3.10.0
packaging==24.0
pandas==2.2.1
parameterized==0.9.0
peft==0.10.0
pillow==10.3.0
piq==0.8.0
platformdirs==4.2.0
portalocker==2.8.2
prettytable==3.10.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
pyre-extensions==0.0.23
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-fid==0.3.0
pytorch-lightning==1.0.8
pytorchvideo==0.1.5
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
resize-right==0.0.2
rich==13.7.1
rotary-embedding-torch==0.3.6
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.3.5
sacred==0.8.5
sacremoses==0.1.1
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
SQLAlchemy==2.0.29
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tb-nightly==2.17.0a20240404
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
test_tube==0.7.5
threadpoolctl==3.4.0
tifffile==2024.2.12
tokenizers==0.19
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torchmetrics==1.3.2
tqdm==4.66.2
transformers==4.40.0
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
vector-quantize-pytorch==1.8.1
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.2
wrapt==1.16.0
x-clip==0.14.4
xformers==0.0.16
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
gdown
timm